Make category clickable
